requirements.txt
absl-py==0.15.0
einops==0.3.2
ftfy
gdown==3.13.1
hydra-core==1.0.5
matplotlib
numpy
pandas==1.4.1
pyrender==0.1.45
pytorch3d==0.3.0
regex
scipy==1.4.1
tensorboard
torch==1.7.1
torchvision==0.8.2
transformers==4.3.2
trimesh==3.9.34